Initial commit of Real Time Statistics (RTS)
RTS is an application I built to demonstrate how to use Riak Core.  It is a
system that maps incoming events to actions.  The actions effect one or many
statistics that are stored by the system.  These statistics can be accessed in
realtime as the events come in.

For example, you could use this application to keep track of keep track of all
user logins across your entire datacenter of servers.  At any time a user login
or logout event could occur.  Depending on how you code your actions you could
track various statistics such as:

1. How many user's are logged in in total?

2. How many machines is user rzezeski logged into?

3. What's the average amount of time spent on a machine by user rzezeski?

4. How many users are logged onto machine A?

This is a bit of a boring example but you could extend it to an entire array
of msgs that might be generated by a machine.  You could keep track of
segfaults, incorrect password attempts, web server hits, cache misses, whatever.